refactor: refactor config and create vanilla files BREAKING CHANGE: new webapp configuration
