+++ doc/README.md
@@ -1,13 +1,14 @@
 # Webapp configuration
 
-- [App instance configuration](appInstance.md)
-- [App behavior configuration](appConfiguration.md)
+- [App global configuration](globalConfiguration.md)
+- [PowerBI configuration](powerBI.md)
 - [Profiles configuration](profiles.md)
 - [Azure functions configuration](apiAzureFunctions.md)
-- [Dashboards](dashboards.md)
+- [Application Insights configuration](applicationInsights.md)
 - [Scenario Parameters configuration](scenarioParametersConfiguration.md)
 - [Scenario Parameters advanced customization](scenarioParametersCustomization.md)
 - [Theme Customization](themeCustomization.md)
 - [About Content Customization](aboutContentCustomization.md)
-- [Custom metrics](customMetrics.md)
 - [Instance visualization](instanceVisualization.md)
+- [Languages configuration](languages.md)
+- [HelpMenu Configuration](helpMenuConfiguration.md)
