--- doc/appInstance.md
@@ -1,47 +1,0 @@
-# App instance configuration
-
-## Edit configuration files
-
-The file [src/config/AppInstance.js](../src/config/AppInstance.js) defines constants specific to your application
-instance: you will need to change these values when creating a new staging or prod version of your application.
-
-- **AZURE_TENANT_ID** is the identifier of the Azure tenant your application is registered in.\
-  You can find it in _Azure portal > App registration > Overview_.
-- **APP_REGISTRATION_CLIENT_ID** is the identifier of your app registration.\
-  You can find it in _Azure portal > App registration > Overview_.
-- **COSMOTECH_API_SCOPE** is the Cosmo Tech API Platform scope (e.g. http://dev.api.cosmotech.com/platform).
-  If you don't know which one you should use, please contact your Cosmo Tech administrator.
-- **APPLICATION_INSIGHTS_INSTRUMENTATION_KEY** is the instrumentation key generated by _Azure Application Insights_ for
-  your application.\
-  You can find it in _Azure portal > Application Insights > Overview_.\
-  To enable Application Insights in your static web app, please read section "Application Insights" below.
-- **DEFAULT_BASE_PATH** is the URL to the Cosmo Tech API you want to use (e.g. https://dev.api.cosmotech.com).\
-  If you don't know which one you should use, please contact your Cosmo Tech administrator.
-- **ORGANIZATION_ID** is the identifier of your organization in the Cosmo Tech portal.\
-  If you don't know its value, please contact your Cosmo Tech administrator.
-- **WORKSPACE_ID** is the identifier of the workspace you created with the Cosmo Tech API.\
-  If you don't know its value, please contact your Cosmo Tech administrator.
-
-## Application Insights
-
-Steps to enable Application Insights in the webapp:
-
-- in the Azure portal:
-  - open the page of your Static Web Apps instance
-  - select "Application Insights" in the left panel
-  - select "Yes" next to "Enable Application Insights" (see note below if you can't enable Application Insights)
-  - click on "Save"
-- then, edit the configuration file _src/config/AppInstance.js_ in the source code of your webapp:
-  - replace the value of **APPLICATION_INSIGHTS_INSTRUMENTATION_KEY** with the instrumentation key that has been
-    generated for your application
-  - set **ENABLE_APPLICATION_INSIGHTS** to true (defined in [src/config/AppConfiguration.js](../src/config/AppConfiguration.js))
-
-> Note: in the Azure portal, if you can't enable Application Insights and encounter an error message such as "_App
-> Insights is only applicable to Static Web Apps with at least one function_", please **check that the value of
-> "api_location" parameter is set to "api"** in your GitHub Actions workflow file, in the .github/workflows folder.
-
-### Power BI
-
-Parameter to define Power BI usages:
-
-- **POWER_BI_WORKSPACE_ID** if _USE_POWER_BI_WITH_USER_CREDENTIALS_ (defined in [src/config/AppInstance.js](../src/config/AppInstance.js)) is set to `true`, it represents the Power BI workspace ID used
