+++ src/config/ApplicationInsights.js
@@ -1,0 +1,6 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// Whether to enable Application Insights events tracking
+export const ENABLE_APPLICATION_INSIGHTS = true;
+export const APPLICATION_INSIGHTS_INSTRUMENTATION_KEY = 'f198a852-15d8-48da-9cd0-1e6d18708489';
