+++ src/config/GlobalConfiguration.vanilla.js
@@ -1,0 +1,17 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// To configure your application instance see
+// https://github.com/Cosmo-Tech/azure-sample-webapp/blob/main/doc/globalConfiguration.md
+
+// Azure App registration details
+export const AZURE_TENANT_ID = '';
+export const APP_REGISTRATION_CLIENT_ID = '';
+
+// API information
+export const COSMOTECH_API_SCOPE = '';
+export const DEFAULT_BASE_PATH = '';
+
+// Cosmo solution identifiers
+export const ORGANIZATION_ID = '';
+export const WORKSPACE_ID = '';
