+++ src/config/HelpMenuConfiguration.js
@@ -1,0 +1,14 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// App version
+export const APP_VERSION = process.env.REACT_APP_VERSION;
+
+// Support page url
+export const SUPPORT_URL = 'https://support.cosmotech.com';
+
+// Cosmotech website url
+export const COSMOTECH_URL = 'https://cosmotech.com';
+
+// Documentation url
+export const DOCUMENTATION_URL = 'https://portal.cosmotech.com/resources/platform-resources/platform-help';
