+++ src/config/HelpMenuConfiguration.vanilla.js
@@ -1,0 +1,17 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// To configure help menu of your application see
+// https://github.com/Cosmo-Tech/azure-sample-webapp/blob/main/doc/helpMenuConfiguration.md
+
+// App version
+export const APP_VERSION = process.env.REACT_APP_VERSION;
+
+// Support page url
+export const SUPPORT_URL = 'https://support.cosmotech.com';
+
+// Cosmotech website url
+export const COSMOTECH_URL = 'https://cosmotech.com';
+
+// Documentation url
+export const DOCUMENTATION_URL = '';
