+++ src/config/Languages.js
@@ -1,0 +1,9 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// App translation configuration
+export const LANGUAGES = {
+  en: 'English',
+  fr: 'Français',
+};
+export const FALLBACK_LANGUAGE = 'en'; // Language used if detection fails
