+++ src/config/Languages.vanilla.js
@@ -1,0 +1,12 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// To configure supported languages in your webapp see
+// https://github.com/Cosmo-Tech/azure-sample-webapp/blob/main/doc/languages.md
+
+// App translation configuration
+export const LANGUAGES = {
+  en: 'English',
+  fr: 'Français',
+};
+export const FALLBACK_LANGUAGE = 'en'; // Language used if detection fails
