+++ src/config/PowerBI.vanilla.js
@@ -1,0 +1,25 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// To configure PowerBI component see
+// https://github.com/Cosmo-Tech/azure-sample-webapp/blob/main/doc/powerBI.md
+
+// Power BI information
+export const POWER_BI_WORKSPACE_ID = '';
+
+// Power BI embedding mode
+export const USE_POWER_BI_WITH_USER_CREDENTIALS = false;
+export const SCENARIO_VIEW_IFRAME_DISPLAY_RATIO = 1580 / 350;
+export const DASHBOARDS_VIEW_IFRAME_DISPLAY_RATIO = 1280 / 795;
+
+// Dashboards configuration
+// For further information about settings or filters see:
+// https://github.com/microsoft/powerbi-client-react
+// based on
+// https://github.com/microsoft/PowerBI-JavaScript
+// using
+// https://github.com/microsoft/powerbi-models
+
+export const SCENARIO_DASHBOARD_CONFIG = {};
+
+export const DASHBOARDS_LIST_CONFIG = [];
