+++ src/config/Profiles.vanilla.js
@@ -1,0 +1,10 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// To configure profiles see:
+// https://github.com/Cosmo-Tech/azure-sample-webapp/blob/main/doc/profiles.md
+// list of existing permissions:
+// src/services/config/Permissions.js
+
+export const APP_ROLES = {};
+export const PROFILES = {};
