+++ src/config/ScenarioParameters.js
@@ -293,6 +293,16 @@ const RUN_TEMPLATES = {
   },
 };
 
+// Additional parameters to put in scenario parameters
+export const ADD_SCENARIO_NAME_PARAMETER = false;
+export const ADD_SCENARIO_ID_PARAMETER = false;
+export const ADD_SCENARIO_LAST_RUN_ID_PARAMETER = false;
+export const ADD_SCENARIO_PARENT_ID_PARAMETER = false;
+export const ADD_SCENARIO_PARENT_LAST_RUN_ID_PARAMETER = false;
+export const ADD_SCENARIO_MASTER_ID_PARAMETER = false;
+export const ADD_SCENARIO_MASTER_LAST_RUN_ID_PARAMETER = false;
+export const ADD_SCENARIO_RUN_TEMPLATE_NAME_PARAMETER = false;
+
 export const SCENARIO_PARAMETERS_CONFIG = {
   parameters: PARAMETERS,
   parametersGroups: PARAMETERS_GROUPS,
