+++ src/services/config/ApplicationInsights.js
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { DistributedTracingModes } from '@microsoft/applicationinsights-web';
-import { APPLICATION_INSIGHTS_INSTRUMENTATION_KEY } from '../../config/AppInstance.js';
+import { APPLICATION_INSIGHTS_INSTRUMENTATION_KEY } from '../../config/ApplicationInsights';
 
 // Application Insight configuration
 export const APPLICATION_INSIGHTS_CONFIG = {
