+++ src/services/config/Auth.js
@@ -3,7 +3,7 @@
 
 import { Auth, AuthDev } from '@cosmotech/core';
 import { AuthMSAL } from '@cosmotech/azure';
-import { APP_REGISTRATION_CLIENT_ID, AZURE_TENANT_ID, COSMOTECH_API_SCOPE } from '../../config/AppInstance.js';
+import { APP_REGISTRATION_CLIENT_ID, AZURE_TENANT_ID, COSMOTECH_API_SCOPE } from '../../config/GlobalConfiguration.js';
 
 export const POWER_BI_API_DEFAULT_SCOPE = 'https://analysis.windows.net/powerbi/api/.default';
