+++ src/services/config/FunctionalConstants.js
@@ -1,0 +1,14 @@
+// Copyright (c) Cosmo Tech.
+// Licensed under the MIT license.
+
+// Duration of inactivity after which users are automatically logged out (minutes)
+export const SESSION_INACTIVITY_TIMEOUT = 30;
+
+// Polling delay to update running scenario status (milliseconds)
+export const SCENARIO_STATUS_POLLING_DELAY = 10000;
+
+// Polling delay to update Power BI information (milliseconds)
+export const POWER_BI_INFO_POLLING_DELAY = 10000;
+
+// Type of logs to download for a scenario run (SIMPLE_LOGS or CUMULATED_LOGS)
+export const SCENARIO_RUN_LOG_TYPE = 'CUMULATED_LOGS';
