+++ src/services/config/i18next.js
@@ -5,7 +5,7 @@ import i18next from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import { initReactI18next } from 'react-i18next';
 import HttpApi from 'i18next-http-backend';
-import { LANGUAGES, FALLBACK_LANGUAGE } from '../../config/AppConfiguration';
+import { LANGUAGES, FALLBACK_LANGUAGE } from '../../config/Languages';
 
 const I18N_NAMESPACE = 'translation';
