+++ src/state/sagas/datasets/FindAllDatasets/FindAllDatasets.js
@@ -3,7 +3,7 @@
 
 import { call, put, takeEvery } from 'redux-saga/effects';
 import { DATASET_ACTIONS_KEY } from '../../../commons/DatasetConstants';
-import { ORGANIZATION_ID } from '../../../../config/AppInstance';
+import { ORGANIZATION_ID } from '../../../../config/GlobalConfiguration';
 import { Api } from '../../../../services/config/Api';
 import { STATUSES } from '../../../commons/Constants';
