+++ src/state/sagas/powerbi/GetPowerBIEmbedInfo/GetPowerBIEmbedInfoData.js
@@ -5,7 +5,8 @@ import { delay, put, takeEvery } from 'redux-saga/effects';
 import { GET_EMBED_INFO_URL, POWER_BI_ACTIONS_KEY } from '../../../commons/PowerBIConstants';
 import { STATUSES } from '../../../commons/Constants';
 import { clientApi } from '../../../../services/ClientApi';
-import { POWER_BI_INFO_POLLING_DELAY, USE_POWER_BI_WITH_USER_CREDENTIALS } from '../../../../config/AppConfiguration';
+import { USE_POWER_BI_WITH_USER_CREDENTIALS } from '../../../../config/PowerBI';
+import { POWER_BI_INFO_POLLING_DELAY } from '../../../../services/config/FunctionalConstants';
 import { PowerBIService } from '../../../../services/powerbi/PowerBIService';
 
 const noAccess = {
