+++ src/state/sagas/scenario/FindAllScenarios/FindAllScenariosData.js
@@ -3,7 +3,7 @@
 
 import { put, takeEvery, call } from 'redux-saga/effects';
 import { SCENARIO_ACTIONS_KEY } from '../../../commons/ScenarioConstants';
-import { ORGANIZATION_ID } from '../../../../config/AppInstance';
+import { ORGANIZATION_ID } from '../../../../config/GlobalConfiguration';
 import { Api } from '../../../../services/config/Api';
 import { STATUSES } from '../../../commons/Constants';
