+++ src/views/ScenarioManager/ScenarioManager.js
@@ -7,7 +7,7 @@ import PropTypes from 'prop-types';
 import { ScenarioUtils } from '@cosmotech/core';
 import { makeStyles } from '@material-ui/core';
 import { ScenarioManagerTreeList } from '@cosmotech/ui';
-import { WORKSPACE_ID } from '../../config/AppInstance';
+import { WORKSPACE_ID } from '../../config/GlobalConfiguration';
 import { useTranslation } from 'react-i18next';
 import { PERMISSIONS } from '../../services/config/Permissions';
 import { PermissionsGate } from '../../components/PermissionsGate';
